+++ src/main/java/paratrip/paratrip/paragliding/repository/ParaglidingRepository.java
@@ -1,13 +1,13 @@
-package paratrip.paratrip.home.paragliding.repository;
+package paratrip.paratrip.paragliding.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import paratrip.paratrip.home.paragliding.entity.Paragliding;
-import paratrip.paratrip.home.paragliding.entity.Region;
+import paratrip.paratrip.paragliding.entity.Paragliding;
+import paratrip.paratrip.paragliding.entity.Region;
 
 import java.util.List;
 
 /**
- * packageName    : paratrip.paratrip.home.paragliding.repository
+ * packageName    : paratrip.paratrip.paragliding.repository
  * fileName       : ParaglidingRepository
  * author         : tlswl
  * date           : 2024-09-09
