+++ src/main/java/paratrip/paratrip/paragliding/service/ParaglidingService.java
@@ -1,21 +1,21 @@
-package paratrip.paratrip.home.paragliding.service;
+package paratrip.paratrip.paragliding.service;
 
 import com.amazonaws.services.s3.model.Region;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import paratrip.paratrip.home.paragliding.dto.request.ParaglidingRequestDto;
-import paratrip.paratrip.home.paragliding.dto.response.ParaglidingResponseDto;
-import paratrip.paratrip.home.paragliding.dto.response.RegionResponseDto;
-import paratrip.paratrip.home.paragliding.entity.Paragliding;
-import paratrip.paratrip.home.paragliding.repository.ParaglidingRepository;
-import paratrip.paratrip.home.paragliding.util.ParaglidingUtils;
+import paratrip.paratrip.paragliding.dto.request.ParaglidingRequestDto;
+import paratrip.paratrip.paragliding.dto.response.ParaglidingResponseDto;
+import paratrip.paratrip.paragliding.dto.response.RegionResponseDto;
+import paratrip.paratrip.paragliding.entity.Paragliding;
+import paratrip.paratrip.paragliding.repository.ParaglidingRepository;
+import paratrip.paratrip.paragliding.util.ParaglidingUtils;
 
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 /**
- * packageName    : paratrip.paratrip.home.paragliding.service
+ * packageName    : paratrip.paratrip.paragliding.service
  * fileName       : ParaglidingService
  * author         : tlswl
  * date           : 2024-09-09
