+++ src/main/java/paratrip/paratrip/paragliding/util/ParaglidingUtils.java
@@ -1,14 +1,14 @@
-package paratrip.paratrip.home.paragliding.util;
+package paratrip.paratrip.paragliding.util;
 
 import org.springframework.stereotype.Component;
-import paratrip.paratrip.home.paragliding.dto.response.ParaglidingResponseDto;
-import paratrip.paratrip.home.paragliding.entity.Paragliding;
+import paratrip.paratrip.paragliding.dto.response.ParaglidingResponseDto;
+import paratrip.paratrip.paragliding.entity.Paragliding;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 /**
- * packageName    : paratrip.paratrip.home.paragliding.util
+ * packageName    : paratrip.paratrip.paragliding.util
  * fileName       : ParaglidingUtils
  * author         : tlswl
  * date           : 2024-09-09
